requirements.txt
exchangelib==5.4.3
PyYAML==6.0.2
